pyproject.toml
[tool.poetry]
name = "qsp"
version = "0.1.0"
description = ""
authors = ["Guilherme <gui.dias.alves20@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
textual = "^0.29.0"
pyyaml = "^6.0"
jinja2 = "^3.1.2"
marshmallow-dataclass = "^8.5.14"
jinja2schema = "^0.1.4"
django = "^4.2.5"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"
black = "^23.3.0"
taskipy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
qsp = "qsp.qsp_cli:app"

[tool.taskipy.tasks]
lint = "black --check --diff . && isort --check --diff ."